Major bugs fixed in AnalyticsCounter.java
Bug 1 - AnalyticsCounter.java
Two variants for the same use: headacheCount (global var, line 8) and headCount (local var, line18).
That is why the headacheCount doesn't work.

Modifications:
- delete "headCount" var declaration line 18,
- replace "headCount" with "headacheCount" line 23: ( headacheCount++; )
- replace "headCount" with "headacheCount" line 24: ( System.out.println("number of headaches: " + headacheCount); )

With this bugfix, the count of headache in result.out is right (3).

Bug 2 - AnalyticsCounter.java - line 26
Looking for "rush" instead of "rash"
That is why the rashCount doesn't work.

Change: wrong search pattern "rush" replaced by "rash" ( else if (line.equals("rash")) { ).

With this bugfix, the count of rash in result.out is right (4).